feat: add bridge themes
